requirements.txt
# Web Server
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
requests

# AI Inference (CPU Optimized)
torch --index-url https://download.pytorch.org/whl/cpu
numpy
scipy

# Signal Processing Dependencies
PyWavelets
filterpy